chore: Break up widgetBuilder.spec.tsx into multiple test files
We parallilze tests by test files in ci. This file has 120 tests
so breaking it up into multiple test files so we don't have such
a long running test.
