+++ static/app/views/dashboardsV2/widgetBuilder/widgetBuilder.spec.tsx
@@ -127,10 +127,7 @@ describe('WidgetBuilder', function () {
   let eventsStatsMock: jest.Mock | undefined;
   let eventsv2Mock: jest.Mock | undefined;
   let eventsMock: jest.Mock | undefined;
-  let sessionsDataMock: jest.Mock | undefined;
-  let metricsDataMock: jest.Mock | undefined;
   let tagsMock: jest.Mock | undefined;
-  let measurementsMetaMock: jest.Mock | undefined;
 
   beforeEach(function () {
     MockApiClient.addMockResponse({
@@ -212,7 +209,7 @@ describe('WidgetBuilder', function () {
       body: [],
     });
 
-    sessionsDataMock = MockApiClient.addMockResponse({
+    MockApiClient.addMockResponse({
       method: 'GET',
       url: '/organizations/org-slug/sessions/',
       body: TestStubs.SessionsField({
@@ -220,7 +217,7 @@ describe('WidgetBuilder', function () {
       }),
     });
 
-    metricsDataMock = MockApiClient.addMockResponse({
+    MockApiClient.addMockResponse({
       method: 'GET',
       url: '/organizations/org-slug/metrics/data/',
       body: TestStubs.MetricsField({
@@ -234,7 +231,7 @@ describe('WidgetBuilder', function () {
       body: TestStubs.Tags(),
     });
 
-    measurementsMetaMock = MockApiClient.addMockResponse({
+    MockApiClient.addMockResponse({
       url: '/organizations/org-slug/measurements-meta/',
       method: 'GET',
       body: {},
@@ -1382,680 +1379,6 @@ describe('WidgetBuilder', function () {
       expect(alertMock).not.toHaveBeenCalled();
     });
 
-    describe('Sort by selectors', function () {
-      it('renders', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-        });
-
-        expect(await screen.findByText('Sort by a column')).toBeInTheDocument();
-        expect(
-          screen.getByText("Choose one of the columns you've created to sort by.")
-        ).toBeInTheDocument();
-
-        // Selector "sortDirection"
-        expect(screen.getByText('High to low')).toBeInTheDocument();
-        // Selector "sortBy"
-        expect(screen.getAllByText('count()')).toHaveLength(3);
-      });
-
-      it('ordering by column uses field form when selecting orderby', async function () {
-        const widget: Widget = {
-          id: '1',
-          title: 'Test Widget',
-          interval: '5m',
-          displayType: DisplayType.TABLE,
-          queries: [
-            {
-              name: 'errors',
-              conditions: 'event.type:error',
-              fields: ['count()'],
-              aggregates: ['count()'],
-              columns: ['project'],
-              orderby: '-project',
-            },
-          ],
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          dashboard,
-          params: {
-            widgetIndex: '0',
-          },
-        });
-
-        await waitFor(async () => {
-          expect(await screen.findAllByText('project')).toHaveLength(3);
-        });
-
-        await selectEvent.select(screen.getAllByText('project')[2], 'count()');
-
-        await waitFor(() => {
-          expect(eventsv2Mock).toHaveBeenCalledWith(
-            '/organizations/org-slug/eventsv2/',
-            expect.objectContaining({
-              query: expect.objectContaining({
-                sort: ['-count()'],
-              }),
-            })
-          );
-        });
-      });
-
-      it('sortBy defaults to the first field value when changing display type to table', async function () {
-        const widget: Widget = {
-          id: '1',
-          title: 'Errors over time',
-          interval: '5m',
-          displayType: DisplayType.LINE,
-          queries: [
-            {
-              name: 'errors',
-              conditions: 'event.type:error',
-              fields: ['count()', 'count_unique(id)'],
-              aggregates: ['count()', 'count_unique(id)'],
-              columns: [],
-              orderby: '',
-            },
-            {
-              name: 'csp',
-              conditions: 'event.type:csp',
-              fields: ['count()', 'count_unique(id)'],
-              aggregates: ['count()', 'count_unique(id)'],
-              columns: [],
-              orderby: '',
-            },
-          ],
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          dashboard,
-          params: {
-            widgetIndex: '0',
-          },
-        });
-
-        // Click on the displayType selector
-        userEvent.click(await screen.findByText('Line Chart'));
-
-        // Choose the table visualization
-        userEvent.click(screen.getByText('Table'));
-
-        expect(await screen.findByText('Sort by a column')).toBeInTheDocument();
-
-        // Selector "sortDirection"
-        expect(screen.getByText('High to low')).toBeInTheDocument();
-
-        // Selector "sortBy"
-        expect(screen.getAllByText('count()')).toHaveLength(3);
-      });
-
-      it('can update selectors values', async function () {
-        const handleSave = jest.fn();
-
-        const widget: Widget = {
-          id: '1',
-          title: 'Errors over time',
-          interval: '5m',
-          displayType: DisplayType.TABLE,
-          queries: [
-            {
-              name: '',
-              conditions: '',
-              fields: ['count()', 'count_unique(id)'],
-              aggregates: ['count()', 'count_unique(id)'],
-              columns: [],
-              orderby: '-count()',
-            },
-          ],
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          dashboard,
-          onSave: handleSave,
-          params: {
-            widgetIndex: '0',
-          },
-        });
-
-        expect(await screen.findByText('Sort by a column')).toBeInTheDocument();
-
-        // Selector "sortDirection"
-        expect(screen.getByText('High to low')).toBeInTheDocument();
-
-        // Selector "sortBy"
-        expect(screen.getAllByText('count()')).toHaveLength(3);
-
-        await selectEvent.select(screen.getAllByText('count()')[2], 'count_unique(id)');
-
-        // Wait for the Builder update the widget values
-        await waitFor(() => {
-          expect(screen.getAllByText('count()')).toHaveLength(2);
-        });
-
-        // Now count_unique(id) is selected in the "sortBy" selector
-        expect(screen.getAllByText('count_unique(id)')).toHaveLength(2);
-
-        await selectEvent.select(screen.getByText('High to low'), 'Low to high');
-
-        // Saves the widget
-        userEvent.click(screen.getByText('Update Widget'));
-
-        await waitFor(() => {
-          expect(handleSave).toHaveBeenCalledWith([
-            expect.objectContaining({
-              queries: [expect.objectContaining({orderby: 'count_unique(id)'})],
-            }),
-          ]);
-        });
-      });
-
-      it('sortBy defaults to the first field value when coming from discover', async function () {
-        const defaultWidgetQuery = {
-          name: '',
-          fields: ['title', 'count()', 'count_unique(user)', 'epm()', 'count()'],
-          columns: ['title'],
-          aggregates: ['count()', 'count_unique(user)', 'epm()', 'count()'],
-          conditions: 'tag:value',
-          orderby: '',
-        };
-
-        const {router} = renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DISCOVERV2,
-            defaultWidgetQuery: urlEncode(defaultWidgetQuery),
-            displayType: DisplayType.TABLE,
-            defaultTableColumns: ['title', 'count()', 'count_unique(user)', 'epm()'],
-          },
-        });
-
-        expect(await screen.findByText('Sort by a column')).toBeInTheDocument();
-
-        // Selector "sortDirection"
-        expect(screen.getByText('Low to high')).toBeInTheDocument();
-
-        // Selector "sortBy"
-        expect(screen.getAllByText('title')).toHaveLength(2);
-
-        // Saves the widget
-        userEvent.click(screen.getByText('Add Widget'));
-
-        await waitFor(() => {
-          expect(router.push).toHaveBeenCalledWith(
-            expect.objectContaining({
-              query: expect.objectContaining({queryOrderby: 'count()'}),
-            })
-          );
-        });
-      });
-
-      it('sortBy is only visible on tabular visualizations or when there is a groupBy value selected on time-series visualizations', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-        });
-
-        // Sort by shall be visible on table visualization
-        expect(await screen.findByText('Sort by a column')).toBeInTheDocument();
-
-        // Update visualization to be a time-series
-        userEvent.click(screen.getByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-
-        // Time-series visualizations display GroupBy step
-        expect(await screen.findByText('Group your results')).toBeInTheDocument();
-
-        // Do not show sortBy when empty columns (groupBys) are added
-        userEvent.click(screen.getByText('Add Group'));
-        expect(screen.getAllByText('Select group')).toHaveLength(2);
-
-        // SortBy step shall not be visible
-        expect(screen.queryByText('Sort by a y-axis')).not.toBeInTheDocument();
-
-        // Select GroupBy value
-        await selectEvent.select(screen.getAllByText('Select group')[0], 'project');
-
-        // Now that at least one groupBy value is selected, the SortBy step shall be visible
-        expect(screen.getByText('Sort by a y-axis')).toBeInTheDocument();
-
-        // Remove selected GroupBy value
-        userEvent.click(screen.getAllByLabelText('Remove group')[0]);
-
-        // SortBy step shall no longer be visible
-        expect(screen.queryByText('Sort by a y-axis')).not.toBeInTheDocument();
-      });
-
-      it('allows for sorting by a custom equation', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], 'Custom Equation');
-        userEvent.paste(
-          screen.getByPlaceholderText('Enter Equation'),
-          'count_unique(user) * 2'
-        );
-        userEvent.keyboard('{enter}');
-
-        await waitFor(() => {
-          expect(eventsStatsMock).toHaveBeenCalledWith(
-            '/organizations/org-slug/events-stats/',
-            expect.objectContaining({
-              query: expect.objectContaining({
-                field: expect.arrayContaining(['equation|count_unique(user) * 2']),
-                orderby: '-equation[0]',
-              }),
-            })
-          );
-        });
-      }, 10000);
-
-      it('persists the state when toggling between sorting options', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], 'Custom Equation');
-        userEvent.paste(
-          screen.getByPlaceholderText('Enter Equation'),
-          'count_unique(user) * 2'
-        );
-        userEvent.keyboard('{enter}');
-
-        // Switch away from the Custom Equation
-        expect(screen.getByText('project')).toBeInTheDocument();
-        await selectEvent.select(screen.getByText('Custom Equation'), 'project');
-        expect(screen.getAllByText('project')).toHaveLength(2);
-
-        // Switch back, the equation should still be visible
-        await selectEvent.select(screen.getAllByText('project')[1], 'Custom Equation');
-        expect(screen.getByPlaceholderText('Enter Equation')).toHaveValue(
-          'count_unique(user) * 2'
-        );
-      });
-
-      it('persists the state when updating y-axes', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], 'Custom Equation');
-        userEvent.paste(
-          screen.getByPlaceholderText('Enter Equation'),
-          'count_unique(user) * 2'
-        );
-        userEvent.keyboard('{enter}');
-
-        // Add a y-axis
-        userEvent.click(screen.getByText('Add Overlay'));
-
-        // The equation should still be visible
-        expect(screen.getByPlaceholderText('Enter Equation')).toHaveValue(
-          'count_unique(user) * 2'
-        );
-      });
-
-      it('displays the custom equation if the widget has it saved', async function () {
-        const widget: Widget = {
-          id: '1',
-          title: 'Test Widget',
-          interval: '5m',
-          displayType: DisplayType.LINE,
-          queries: [
-            {
-              name: '',
-              conditions: '',
-              fields: ['count()', 'project'],
-              aggregates: ['count()'],
-              columns: ['project'],
-              orderby: '-equation|count_unique(user) * 2',
-            },
-          ],
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-          params: {
-            widgetIndex: '0',
-          },
-          dashboard,
-        });
-
-        expect(await screen.findByPlaceholderText('Enter Equation')).toHaveValue(
-          'count_unique(user) * 2'
-        );
-      });
-
-      it('displays Operators in the input dropdown', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], 'Custom Equation');
-        selectEvent.openMenu(screen.getByPlaceholderText('Enter Equation'));
-
-        expect(screen.getByText('Operators')).toBeInTheDocument();
-        expect(screen.queryByText('Fields')).not.toBeInTheDocument();
-      });
-
-      it('hides Custom Equation input and resets orderby when switching to table', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], 'Custom Equation');
-        userEvent.paste(
-          screen.getByPlaceholderText('Enter Equation'),
-          'count_unique(user) * 2'
-        );
-        userEvent.keyboard('{enter}');
-
-        // Switch the display type to Table
-        userEvent.click(screen.getByText('Line Chart'));
-        userEvent.click(screen.getByText('Table'));
-
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        expect(screen.queryByPlaceholderText('Enter Equation')).not.toBeInTheDocument();
-
-        await waitFor(() => {
-          expect(eventsv2Mock).toHaveBeenCalledWith(
-            '/organizations/org-slug/eventsv2/',
-            expect.objectContaining({
-              query: expect.objectContaining({
-                sort: ['-count()'],
-              }),
-            })
-          );
-        });
-      });
-
-      it('does not show the Custom Equation input if the only y-axis left is an empty equation', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        userEvent.click(screen.getByText('Add an Equation'));
-        userEvent.click(screen.getAllByLabelText('Remove this Y-Axis')[0]);
-
-        expect(screen.queryByPlaceholderText('Enter Equation')).not.toBeInTheDocument();
-      });
-
-      it('persists a sort by a grouping when changing y-axes', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-
-        // Change the sort option to a grouping field, and then change a y-axis
-        await selectEvent.select(screen.getAllByText('count()')[1], 'project');
-        await selectEvent.select(screen.getAllByText('count()')[0], /count_unique/);
-
-        // project should appear in the group by field, as well as the sort field
-        expect(screen.getAllByText('project')).toHaveLength(2);
-      });
-
-      it('persists sort by a y-axis when grouping changes', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        userEvent.click(await screen.findByText('Add Overlay'));
-        await selectEvent.select(screen.getByText('Select group'), 'project');
-
-        // Change the sort by to count_unique
-        await selectEvent.select(screen.getAllByText('count()')[1], /count_unique/);
-
-        // Change the grouping
-        await selectEvent.select(screen.getByText('project'), 'environment');
-
-        // count_unique(user) should still be the sorting field
-        expect(screen.getByText(/count_unique/)).toBeInTheDocument();
-        expect(screen.getByText('user')).toBeInTheDocument();
-      });
-
-      it('does not remove the Custom Equation field if a grouping is updated', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        await selectEvent.select(screen.getAllByText('count()')[1], 'Custom Equation');
-        userEvent.paste(
-          screen.getByPlaceholderText('Enter Equation'),
-          'count_unique(user) * 2'
-        );
-        userEvent.keyboard('{enter}');
-
-        userEvent.click(screen.getByText('Add Group'));
-        expect(screen.getByPlaceholderText('Enter Equation')).toHaveValue(
-          'count_unique(user) * 2'
-        );
-      });
-
-      it.each`
-        directionPrefix | expectedOrderSelection | displayType
-        ${'-'}          | ${'High to low'}       | ${DisplayType.TABLE}
-        ${''}           | ${'Low to high'}       | ${DisplayType.TABLE}
-        ${'-'}          | ${'High to low'}       | ${DisplayType.LINE}
-        ${''}           | ${'Low to high'}       | ${DisplayType.LINE}
-      `(
-        `opens a widget with the '$expectedOrderSelection' sort order when the widget was saved with that direction`,
-        async function ({directionPrefix, expectedOrderSelection}) {
-          const widget: Widget = {
-            id: '1',
-            title: 'Test Widget',
-            interval: '5m',
-            displayType: DisplayType.LINE,
-            queries: [
-              {
-                name: '',
-                conditions: '',
-                fields: ['count_unique(user)'],
-                aggregates: ['count_unique(user)'],
-                columns: ['project'],
-                orderby: `${directionPrefix}count_unique(user)`,
-              },
-            ],
-          };
-
-          const dashboard = mockDashboard({widgets: [widget]});
-
-          renderTestComponent({
-            orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-            dashboard,
-            params: {
-              widgetIndex: '0',
-            },
-          });
-
-          await screen.findByText(expectedOrderSelection);
-        }
-      );
-
-      it('saved widget with aggregate alias as orderby should persist alias when y-axes change', async function () {
-        const widget: Widget = {
-          id: '1',
-          title: 'Test Widget',
-          interval: '5m',
-          displayType: DisplayType.TABLE,
-          queries: [
-            {
-              name: '',
-              conditions: '',
-              fields: ['project', 'count_unique(user)'],
-              aggregates: ['count_unique(user)'],
-              columns: ['project'],
-              orderby: 'count_unique(user)',
-            },
-          ],
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          dashboard,
-          params: {
-            widgetIndex: '0',
-          },
-        });
-
-        await screen.findByText('Sort by a column');
-
-        // Assert for length 2 since one in the table header and one in sort by
-        expect(screen.getAllByText('count_unique(user)')).toHaveLength(2);
-
-        userEvent.click(screen.getByText('Add a Column'));
-
-        // The sort by should still have count_unique(user)
-        await waitFor(() =>
-          expect(screen.getAllByText('count_unique(user)')).toHaveLength(2)
-        );
-      });
-
-      it('will reset the sort field when going from line to table when sorting by a value not in fields', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], /count_unique/);
-
-        userEvent.click(screen.getByText('Line Chart'));
-        userEvent.click(screen.getByText('Table'));
-
-        // 1 for table header, 1 for column selection, and 1 for sorting
-        await waitFor(() => {
-          expect(screen.getAllByText('count()')).toHaveLength(3);
-        });
-      });
-
-      it('equations in y-axis appear in sort by field for grouped timeseries', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-          query: {
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        userEvent.click(await screen.findByText('Add an Equation'));
-        userEvent.paste(screen.getByPlaceholderText('Equation'), 'count() * 100');
-        userEvent.keyboard('{enter}');
-
-        await selectEvent.select(screen.getByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], 'count() * 100');
-      });
-
-      it('does not reset the orderby when ordered by an equation in table', async function () {
-        const widget: Widget = {
-          id: '1',
-          title: 'Errors over time',
-          interval: '5m',
-          displayType: DisplayType.TABLE,
-          queries: [
-            {
-              name: '',
-              conditions: '',
-              fields: [
-                'count()',
-                'count_unique(id)',
-                'equation|count() + count_unique(id)',
-              ],
-              aggregates: [
-                'count()',
-                'count_unique(id)',
-                'equation|count() + count_unique(id)',
-              ],
-              columns: [],
-              orderby: '-equation[0]',
-            },
-          ],
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-
-        renderTestComponent({
-          dashboard,
-          params: {
-            widgetIndex: '0',
-          },
-        });
-
-        await screen.findByText('Sort by a column');
-
-        // 1 in the column selector, 1 in the sort by field
-        expect(screen.getAllByText('count() + count_unique(id)')).toHaveLength(2);
-      });
-    });
-
     describe('Widget creation coming from other verticals', function () {
       it('redirects correctly when creating a new dashboard', async function () {
         const {router} = renderTestComponent({
@@ -2320,546 +1643,6 @@ describe('WidgetBuilder', function () {
         '/organizations/org-slug/events-stats/',
         expectedArgs
       );
-    });
-
-    describe('Issue Widgets', function () {
-      it('sets widgetType to issues', async function () {
-        const handleSave = jest.fn();
-
-        renderTestComponent({onSave: handleSave});
-
-        userEvent.click(
-          await screen.findByText('Issues (States, Assignment, Time, etc.)')
-        );
-        userEvent.click(screen.getByLabelText('Add Widget'));
-
-        await waitFor(() => {
-          expect(handleSave).toHaveBeenCalledWith([
-            expect.objectContaining({
-              title: 'Custom Widget',
-              displayType: DisplayType.TABLE,
-              interval: '5m',
-              widgetType: WidgetType.ISSUE,
-              queries: [
-                {
-                  conditions: '',
-                  fields: ['issue', 'assignee', 'title'],
-                  columns: ['issue', 'assignee', 'title'],
-                  aggregates: [],
-                  fieldAliases: [],
-                  name: '',
-                  orderby: 'date',
-                },
-              ],
-            }),
-          ]);
-        });
-
-        expect(handleSave).toHaveBeenCalledTimes(1);
-      });
-
-      it('render issues dataset disabled when the display type is not set to table', async function () {
-        renderTestComponent({
-          query: {
-            source: DashboardWidgetSource.DISCOVERV2,
-          },
-        });
-
-        userEvent.click(await screen.findByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-        expect(
-          screen.getByRole('radio', {
-            name: 'Select Errors and Transactions',
-          })
-        ).toBeEnabled();
-        expect(
-          screen.getByRole('radio', {
-            name: 'Select Issues (States, Assignment, Time, etc.)',
-          })
-        ).toBeDisabled();
-      });
-
-      it('disables moving and deleting issue column', async function () {
-        renderTestComponent();
-
-        userEvent.click(
-          await screen.findByText('Issues (States, Assignment, Time, etc.)')
-        );
-        expect(screen.getByText('issue')).toBeInTheDocument();
-        expect(screen.getByText('assignee')).toBeInTheDocument();
-        expect(screen.getByText('title')).toBeInTheDocument();
-        expect(screen.getAllByLabelText('Remove column')).toHaveLength(2);
-        expect(screen.getAllByLabelText('Drag to reorder')).toHaveLength(3);
-
-        userEvent.click(screen.getAllByLabelText('Remove column')[1]);
-        userEvent.click(screen.getAllByLabelText('Remove column')[0]);
-
-        expect(screen.getByText('issue')).toBeInTheDocument();
-        expect(screen.queryByText('assignee')).not.toBeInTheDocument();
-        expect(screen.queryByText('title')).not.toBeInTheDocument();
-        expect(screen.queryByLabelText('Remove column')).not.toBeInTheDocument();
-        expect(screen.queryByLabelText('Drag to reorder')).not.toBeInTheDocument();
-      });
-
-      it('issue query does not work on default search bar', async function () {
-        renderTestComponent();
-
-        const input = (await screen.findByPlaceholderText(
-          'Search for events, users, tags, and more'
-        )) as HTMLTextAreaElement;
-        userEvent.paste(input, 'bookmarks', {
-          clipboardData: {getData: () => ''},
-        } as unknown as React.ClipboardEvent<HTMLTextAreaElement>);
-        input.setSelectionRange(9, 9);
-
-        expect(await screen.findByText('No items found')).toBeInTheDocument();
-      });
-
-      it('renders with an issues search bar when selected in dataset selection', async function () {
-        renderTestComponent();
-
-        userEvent.click(
-          await screen.findByText('Issues (States, Assignment, Time, etc.)')
-        );
-
-        const input = (await screen.findByPlaceholderText(
-          'Search for issues, status, assigned, and more'
-        )) as HTMLTextAreaElement;
-        userEvent.paste(input, 'is:', {
-          clipboardData: {getData: () => ''},
-        } as unknown as React.ClipboardEvent<HTMLTextAreaElement>);
-        input.setSelectionRange(3, 3);
-
-        expect(await screen.findByText('resolved')).toBeInTheDocument();
-      });
-
-      it('Update table header values (field alias)', async function () {
-        const handleSave = jest.fn();
-
-        renderTestComponent({
-          onSave: handleSave,
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-        });
-
-        await screen.findByText('Table');
-
-        userEvent.click(screen.getByText('Issues (States, Assignment, Time, etc.)'));
-
-        userEvent.paste(screen.getAllByPlaceholderText('Alias')[0], 'First Alias');
-
-        userEvent.click(screen.getByText('Add Widget'));
-
-        await waitFor(() => {
-          expect(handleSave).toHaveBeenCalledWith([
-            expect.objectContaining({
-              queries: [
-                expect.objectContaining({
-                  fieldAliases: ['First Alias', '', ''],
-                }),
-              ],
-            }),
-          ]);
-        });
-      });
-    });
-
-    describe('Release Widgets', function () {
-      const releaseHealthFeatureFlags = [
-        ...defaultOrgFeatures,
-        'new-widget-builder-experience-design',
-        'dashboards-releases',
-      ];
-
-      it('does not show the Release Health dataset if there is no dashboards-releases flag', async function () {
-        renderTestComponent({
-          orgFeatures: [...defaultOrgFeatures, 'new-widget-builder-experience-design'],
-        });
-
-        expect(await screen.findByText('Errors and Transactions')).toBeInTheDocument();
-        expect(
-          screen.queryByText('Releases (Sessions, Crash rates)')
-        ).not.toBeInTheDocument();
-      });
-
-      it('shows the Release Health dataset if there is the dashboards-releases flag', async function () {
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(await screen.findByText('Errors and Transactions')).toBeInTheDocument();
-        expect(screen.getByText('Releases (Sessions, Crash rates)')).toBeInTheDocument();
-      });
-
-      it('maintains the selected dataset when display type is changed', async function () {
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        expect(screen.getByLabelText(/releases/i)).not.toBeChecked();
-        userEvent.click(screen.getByLabelText(/releases/i));
-        await waitFor(() => expect(screen.getByLabelText(/releases/i)).toBeChecked());
-
-        userEvent.click(screen.getByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-        await waitFor(() => expect(screen.getByLabelText(/releases/i)).toBeChecked());
-      });
-
-      it('displays releases tags', async function () {
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        expect(screen.getByText('crash_free_rate(…)')).toBeInTheDocument();
-        expect(screen.getByText('session')).toBeInTheDocument();
-
-        userEvent.click(screen.getByText('crash_free_rate(…)'));
-        expect(screen.getByText('count_unique(…)')).toBeInTheDocument();
-
-        expect(screen.getByText('release')).toBeInTheDocument();
-        expect(screen.getByText('environment')).toBeInTheDocument();
-        expect(screen.getByText('session.status')).toBeInTheDocument();
-
-        userEvent.click(screen.getByText('count_unique(…)'));
-        expect(screen.getByText('user')).toBeInTheDocument();
-      });
-
-      it('does not display tags as params', async function () {
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        expect(screen.getByText('crash_free_rate(…)')).toBeInTheDocument();
-        await selectEvent.select(
-          screen.getByText('crash_free_rate(…)'),
-          'count_unique(…)'
-        );
-
-        userEvent.click(screen.getByText('user'));
-        expect(screen.queryByText('release')).not.toBeInTheDocument();
-        expect(screen.queryByText('environment')).not.toBeInTheDocument();
-        expect(screen.queryByText('session.status')).not.toBeInTheDocument();
-      });
-
-      it('does not allow sort by when session.status is selected', async function () {
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        expect(screen.getByText('High to low')).toBeEnabled();
-        expect(screen.getByText('crash_free_rate(session)')).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText('Add a Column'));
-        await selectEvent.select(screen.getByText('(Required)'), 'session.status');
-
-        expect(screen.getByRole('textbox', {name: 'Sort direction'})).toBeDisabled();
-        expect(screen.getByRole('textbox', {name: 'Sort by'})).toBeDisabled();
-      });
-
-      it('does not allow sort on tags except release', async function () {
-        jest.useFakeTimers().setSystemTime(new Date('2022-08-02'));
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        expect(screen.getByText('High to low')).toBeEnabled();
-        expect(screen.getByText('crash_free_rate(session)')).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText('Add a Column'));
-        await selectEvent.select(screen.getByText('(Required)'), 'release');
-
-        userEvent.click(screen.getByLabelText('Add a Column'));
-        await selectEvent.select(screen.getByText('(Required)'), 'environment');
-
-        expect(await screen.findByText('Sort by a column')).toBeInTheDocument();
-
-        // Selector "sortDirection"
-        expect(screen.getByText('High to low')).toBeInTheDocument();
-
-        // Selector "sortBy"
-        userEvent.click(screen.getAllByText('crash_free_rate(session)')[1]);
-
-        // release exists in sort by selector
-        expect(screen.getAllByText('release')).toHaveLength(3);
-        // environment does not exist in sort by selector
-        expect(screen.getAllByText('environment')).toHaveLength(2);
-      });
-
-      it('makes the appropriate sessions call', async function () {
-        jest.useFakeTimers().setSystemTime(new Date('2022-08-02'));
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        userEvent.click(screen.getByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-
-        await waitFor(() =>
-          expect(metricsDataMock).toHaveBeenLastCalledWith(
-            `/organizations/org-slug/metrics/data/`,
-            expect.objectContaining({
-              query: expect.objectContaining({
-                environment: [],
-                field: [`session.crash_free_rate`],
-                groupBy: [],
-                interval: '5m',
-                project: [],
-                statsPeriod: '24h',
-              }),
-            })
-          )
-        );
-      });
-
-      it('calls the session endpoint with the right limit', async function () {
-        jest.useFakeTimers().setSystemTime(new Date('2022-08-02'));
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        userEvent.click(screen.getByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-
-        expect(screen.getByText('Limit to 5 results')).toBeInTheDocument();
-
-        await waitFor(() =>
-          expect(metricsDataMock).toHaveBeenLastCalledWith(
-            `/organizations/org-slug/metrics/data/`,
-            expect.objectContaining({
-              query: expect.objectContaining({
-                environment: [],
-                field: ['session.crash_free_rate'],
-                groupBy: ['project_id'],
-                interval: '5m',
-                orderBy: '-session.crash_free_rate',
-                per_page: 5,
-                project: [],
-                statsPeriod: '24h',
-              }),
-            })
-          )
-        );
-      });
-
-      it('calls sessions api when session.status is selected as a groupby', async function () {
-        jest.useFakeTimers().setSystemTime(new Date('2022-08-02'));
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        userEvent.click(screen.getByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-
-        await selectEvent.select(
-          await screen.findByText('Select group'),
-          'session.status'
-        );
-
-        expect(screen.getByText('Limit to 5 results')).toBeInTheDocument();
-
-        await waitFor(() =>
-          expect(sessionsDataMock).toHaveBeenLastCalledWith(
-            `/organizations/org-slug/sessions/`,
-            expect.objectContaining({
-              query: expect.objectContaining({
-                environment: [],
-                field: ['crash_free_rate(session)'],
-                groupBy: ['session.status'],
-                interval: '5m',
-                project: [],
-                statsPeriod: '24h',
-              }),
-            })
-          )
-        );
-      });
-
-      it('displays the correct options for area chart', async function () {
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        expect(
-          await screen.findByText('Releases (Sessions, Crash rates)')
-        ).toBeInTheDocument();
-
-        // change dataset to releases
-        userEvent.click(screen.getByLabelText(/releases/i));
-
-        userEvent.click(screen.getByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-
-        expect(screen.getByText('crash_free_rate(…)')).toBeInTheDocument();
-        expect(screen.getByText(`session`)).toBeInTheDocument();
-
-        userEvent.click(screen.getByText('crash_free_rate(…)'));
-        expect(screen.getByText('count_unique(…)')).toBeInTheDocument();
-
-        userEvent.click(screen.getByText('count_unique(…)'));
-        expect(screen.getByText('user')).toBeInTheDocument();
-      });
-
-      it('sets widgetType to release', async function () {
-        jest.useFakeTimers().setSystemTime(new Date('2022-08-02'));
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        userEvent.click(await screen.findByText('Releases (Sessions, Crash rates)'));
-
-        expect(metricsDataMock).toHaveBeenCalled();
-        expect(screen.getByLabelText(/Releases/i)).toBeChecked();
-      });
-
-      it('does not display "add an equation" button', async function () {
-        const widget: Widget = {
-          title: 'Release Widget',
-          displayType: DisplayType.TABLE,
-          widgetType: WidgetType.RELEASE,
-          queries: [
-            {
-              name: 'errors',
-              conditions: 'event.type:error',
-              fields: ['sdk.name', 'count()'],
-              columns: ['sdk.name'],
-              aggregates: ['count()'],
-              orderby: '-sdk.name',
-            },
-          ],
-          interval: '1d',
-          id: '1',
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-          dashboard,
-          params: {
-            widgetIndex: '0',
-          },
-        });
-
-        // Select line chart display
-        userEvent.click(await screen.findByText('Table'));
-        userEvent.click(screen.getByText('Line Chart'));
-
-        await waitFor(() =>
-          expect(screen.queryByLabelText('Add an Equation')).not.toBeInTheDocument()
-        );
-      });
-
-      it('render release dataset disabled when the display type is world map', async function () {
-        renderTestComponent({
-          query: {
-            source: DashboardWidgetSource.DISCOVERV2,
-          },
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        userEvent.click(await screen.findByText('Table'));
-        userEvent.click(screen.getByText('World Map'));
-
-        await waitFor(() => expect(screen.getByLabelText(/Releases/)).toBeDisabled());
-
-        expect(
-          screen.getByRole('radio', {
-            name: 'Select Errors and Transactions',
-          })
-        ).toBeEnabled();
-        expect(
-          screen.getByRole('radio', {
-            name: 'Select Issues (States, Assignment, Time, etc.)',
-          })
-        ).toBeDisabled();
-      });
-
-      it('renders with a release search bar', async function () {
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        userEvent.type(
-          await screen.findByPlaceholderText('Search for events, users, tags, and more'),
-          'session.status:'
-        );
-
-        await waitFor(() => {
-          expect(screen.getByText("isn't supported here.")).toBeInTheDocument();
-        });
-
-        userEvent.click(screen.getByText('Releases (Sessions, Crash rates)'));
-        userEvent.click(
-          screen.getByPlaceholderText(
-            'Search for release version, session status, and more'
-          )
-        );
-        expect(await screen.findByText('environment')).toBeInTheDocument();
-        expect(screen.getByText('project')).toBeInTheDocument();
-        expect(screen.getByText('release')).toBeInTheDocument();
-      });
-
-      it('adds a function when the only column chosen in a table is a tag', async function () {
-        jest.useFakeTimers().setSystemTime(new Date('2022-08-02'));
-        renderTestComponent({
-          orgFeatures: releaseHealthFeatureFlags,
-        });
-
-        userEvent.click(await screen.findByText('Releases (Sessions, Crash rates)'));
-
-        await selectEvent.select(screen.getByText('crash_free_rate(…)'), 'environment');
-
-        // 1 in the table header, 1 in the column selector, and 1 in the sort by
-        expect(screen.getAllByText(/crash_free_rate/)).toHaveLength(3);
-        expect(screen.getAllByText('environment')).toHaveLength(2);
-      });
     });
 
     describe('Widget Library', function () {
@@ -3251,99 +2034,6 @@ describe('WidgetBuilder', function () {
       });
     });
 
-    describe('Sort by selectors', function () {
-      it('ordering by column uses field form when selecting orderby', async function () {
-        const widget: Widget = {
-          id: '1',
-          title: 'Test Widget',
-          interval: '5m',
-          displayType: DisplayType.TABLE,
-          queries: [
-            {
-              name: 'errors',
-              conditions: 'event.type:error',
-              fields: ['count()'],
-              aggregates: ['count()'],
-              columns: ['project'],
-              orderby: '-project',
-            },
-          ],
-        };
-
-        const dashboard = mockDashboard({widgets: [widget]});
-
-        renderTestComponent({
-          orgFeatures: [
-            ...defaultOrgFeatures,
-            'new-widget-builder-experience-design',
-            'discover-frontend-use-events-endpoint',
-          ],
-          dashboard,
-          params: {
-            widgetIndex: '0',
-          },
-        });
-
-        await waitFor(async () => {
-          expect(await screen.findAllByText('project')).toHaveLength(3);
-        });
-
-        await selectEvent.select(screen.getAllByText('project')[2], 'count()');
-
-        await waitFor(() => {
-          expect(eventsMock).toHaveBeenCalledWith(
-            '/organizations/org-slug/events/',
-            expect.objectContaining({
-              query: expect.objectContaining({
-                sort: ['-count()'],
-              }),
-            })
-          );
-        });
-      });
-
-      it('hides Custom Equation input and resets orderby when switching to table', async function () {
-        renderTestComponent({
-          orgFeatures: [
-            ...defaultOrgFeatures,
-            'new-widget-builder-experience-design',
-            'discover-frontend-use-events-endpoint',
-          ],
-          query: {
-            source: DashboardWidgetSource.DASHBOARDS,
-            displayType: DisplayType.LINE,
-          },
-        });
-
-        await selectEvent.select(await screen.findByText('Select group'), 'project');
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        await selectEvent.select(screen.getAllByText('count()')[1], 'Custom Equation');
-        userEvent.paste(
-          screen.getByPlaceholderText('Enter Equation'),
-          'count_unique(user) * 2'
-        );
-        userEvent.keyboard('{enter}');
-
-        // Switch the display type to Table
-        userEvent.click(screen.getByText('Line Chart'));
-        userEvent.click(screen.getByText('Table'));
-
-        expect(screen.getAllByText('count()')).toHaveLength(2);
-        expect(screen.queryByPlaceholderText('Enter Equation')).not.toBeInTheDocument();
-
-        await waitFor(() => {
-          expect(eventsMock).toHaveBeenCalledWith(
-            '/organizations/org-slug/events/',
-            expect.objectContaining({
-              query: expect.objectContaining({
-                sort: ['-count()'],
-              }),
-            })
-          );
-        });
-      });
-    });
-
     describe('Widget creation coming from other verticals', function () {
       it('shows the correct orderby when switching from a line chart to table', async function () {
         const defaultWidgetQuery = {
@@ -3386,479 +2076,6 @@ describe('WidgetBuilder', function () {
             })
           );
         });
-      });
-    });
-
-    describe('Custom Performance Metrics', function () {
-      it('can choose a custom measurement', async function () {
-        measurementsMetaMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/measurements-meta/',
-          method: 'GET',
-          body: {'measurements.custom.measurement': {functions: ['p99']}},
-        });
-
-        eventsMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/events/',
-          method: 'GET',
-          statusCode: 200,
-          body: {
-            meta: {
-              fields: {'p99(measurements.total.db.calls)': 'duration'},
-              isMetricsData: true,
-            },
-            data: [{'p99(measurements.total.db.calls)': 10}],
-          },
-        });
-
-        const {router} = renderTestComponent({
-          query: {source: DashboardWidgetSource.DISCOVERV2},
-          dashboard: testDashboard,
-          orgFeatures: [...defaultOrgFeatures, 'discover-frontend-use-events-endpoint'],
-        });
-
-        expect(await screen.findAllByText('Custom Widget')).toHaveLength(2);
-
-        // 1 in the table header, 1 in the column selector, 1 in the sort field
-        const countFields = screen.getAllByText('count()');
-        expect(countFields).toHaveLength(3);
-
-        await selectEvent.select(countFields[1], ['p99(…)']);
-        await selectEvent.select(screen.getByText('transaction.duration'), [
-          'measurements.custom.measurement',
-        ]);
-
-        userEvent.click(screen.getByText('Add Widget'));
-
-        await waitFor(() => {
-          expect(router.push).toHaveBeenCalledWith(
-            expect.objectContaining({
-              pathname: '/organizations/org-slug/dashboard/2/',
-              query: {
-                displayType: 'table',
-                interval: '5m',
-                title: 'Custom Widget',
-                queryNames: [''],
-                queryConditions: [''],
-                queryFields: ['p99(measurements.custom.measurement)'],
-                queryOrderby: '-p99(measurements.custom.measurement)',
-                start: null,
-                end: null,
-                statsPeriod: '24h',
-                utc: null,
-                project: [],
-                environment: [],
-              },
-            })
-          );
-        });
-      });
-
-      it('raises an alert banner and disables saving widget if widget result is not metrics data and widget is using custom measurements', async function () {
-        eventsMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/events/',
-          method: 'GET',
-          statusCode: 200,
-          body: {
-            meta: {
-              fields: {'p99(measurements.custom.measurement)': 'duration'},
-              isMetricsData: false,
-            },
-            data: [{'p99(measurements.custom.measurement)': 10}],
-          },
-        });
-
-        const defaultWidgetQuery = {
-          name: '',
-          fields: ['p99(measurements.custom.measurement)'],
-          columns: [],
-          aggregates: ['p99(measurements.custom.measurement)'],
-          conditions: 'user:test.user@sentry.io',
-          orderby: '',
-        };
-
-        const defaultTableColumns = ['p99(measurements.custom.measurement)'];
-
-        renderTestComponent({
-          query: {
-            source: DashboardWidgetSource.DISCOVERV2,
-            defaultWidgetQuery: urlEncode(defaultWidgetQuery),
-            displayType: DisplayType.TABLE,
-            defaultTableColumns,
-          },
-          orgFeatures: [
-            ...defaultOrgFeatures,
-            'discover-frontend-use-events-endpoint',
-            'dashboards-mep',
-          ],
-        });
-
-        await waitFor(() => {
-          expect(measurementsMetaMock).toHaveBeenCalled();
-        });
-
-        await waitFor(() => {
-          expect(eventsMock).toHaveBeenCalled();
-        });
-
-        screen.getByText('You have inputs that are incompatible with');
-        expect(screen.getByText('Add Widget').closest('button')).toBeDisabled();
-      });
-
-      it('raises an alert banner if widget result is not metrics data', async function () {
-        eventsMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/events/',
-          method: 'GET',
-          statusCode: 200,
-          body: {
-            meta: {
-              fields: {'p99(measurements.lcp)': 'duration'},
-              isMetricsData: false,
-            },
-            data: [{'p99(measurements.lcp)': 10}],
-          },
-        });
-
-        const defaultWidgetQuery = {
-          name: '',
-          fields: ['p99(measurements.lcp)'],
-          columns: [],
-          aggregates: ['p99(measurements.lcp)'],
-          conditions: 'user:test.user@sentry.io',
-          orderby: '',
-        };
-
-        const defaultTableColumns = ['p99(measurements.lcp)'];
-
-        renderTestComponent({
-          query: {
-            source: DashboardWidgetSource.DISCOVERV2,
-            defaultWidgetQuery: urlEncode(defaultWidgetQuery),
-            displayType: DisplayType.TABLE,
-            defaultTableColumns,
-          },
-          orgFeatures: [
-            ...defaultOrgFeatures,
-            'discover-frontend-use-events-endpoint',
-            'dashboards-mep',
-          ],
-        });
-
-        await waitFor(() => {
-          expect(measurementsMetaMock).toHaveBeenCalled();
-        });
-
-        await waitFor(() => {
-          expect(eventsMock).toHaveBeenCalled();
-        });
-
-        screen.getByText('Your selection is only applicable to');
-      });
-
-      it('does not raise an alert banner if widget result is not metrics data but widget contains error fields', async function () {
-        eventsMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/events/',
-          method: 'GET',
-          statusCode: 200,
-          body: {
-            meta: {
-              fields: {'p99(measurements.lcp)': 'duration'},
-              isMetricsData: false,
-            },
-            data: [{'p99(measurements.lcp)': 10}],
-          },
-        });
-
-        const defaultWidgetQuery = {
-          name: '',
-          fields: ['p99(measurements.lcp)'],
-          columns: ['error.handled'],
-          aggregates: ['p99(measurements.lcp)'],
-          conditions: 'user:test.user@sentry.io',
-          orderby: '',
-        };
-
-        const defaultTableColumns = ['p99(measurements.lcp)'];
-
-        renderTestComponent({
-          query: {
-            source: DashboardWidgetSource.DISCOVERV2,
-            defaultWidgetQuery: urlEncode(defaultWidgetQuery),
-            displayType: DisplayType.TABLE,
-            defaultTableColumns,
-          },
-          orgFeatures: [
-            ...defaultOrgFeatures,
-            'discover-frontend-use-events-endpoint',
-            'dashboards-mep',
-          ],
-        });
-
-        await waitFor(() => {
-          expect(measurementsMetaMock).toHaveBeenCalled();
-        });
-
-        await waitFor(() => {
-          expect(eventsMock).toHaveBeenCalled();
-        });
-
-        expect(
-          screen.queryByText('Your selection is only applicable to')
-        ).not.toBeInTheDocument();
-      });
-
-      it('only displays custom measurements in supported functions', async function () {
-        measurementsMetaMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/measurements-meta/',
-          method: 'GET',
-          body: {
-            'measurements.custom.measurement': {functions: ['p99']},
-            'measurements.another.custom.measurement': {functions: ['p95']},
-          },
-        });
-
-        renderTestComponent({
-          query: {source: DashboardWidgetSource.DISCOVERV2},
-          dashboard: testDashboard,
-          orgFeatures: [...defaultOrgFeatures, 'discover-frontend-use-events-endpoint'],
-        });
-
-        expect(await screen.findAllByText('Custom Widget')).toHaveLength(2);
-
-        await selectEvent.select(screen.getAllByText('count()')[1], ['p99(…)']);
-        userEvent.click(screen.getByText('transaction.duration'));
-        screen.getByText('measurements.custom.measurement');
-        expect(
-          screen.queryByText('measurements.another.custom.measurement')
-        ).not.toBeInTheDocument();
-        await selectEvent.select(screen.getAllByText('p99(…)')[0], ['p95(…)']);
-        userEvent.click(screen.getByText('transaction.duration'));
-        screen.getByText('measurements.another.custom.measurement');
-        expect(
-          screen.queryByText('measurements.custom.measurement')
-        ).not.toBeInTheDocument();
-      });
-
-      it('renders custom performance metric using duration units from events meta', async function () {
-        eventsMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/events/',
-          method: 'GET',
-          statusCode: 200,
-          body: {
-            meta: {
-              fields: {'p99(measurements.custom.measurement)': 'duration'},
-              isMetricsData: true,
-              units: {'p99(measurements.custom.measurement)': 'hour'},
-            },
-            data: [{'p99(measurements.custom.measurement)': 12}],
-          },
-        });
-
-        renderTestComponent({
-          query: {source: DashboardWidgetSource.DISCOVERV2},
-          dashboard: {
-            ...testDashboard,
-            widgets: [
-              {
-                title: 'Custom Measurement Widget',
-                interval: '1d',
-                id: '1',
-                widgetType: WidgetType.DISCOVER,
-                displayType: DisplayType.TABLE,
-                queries: [
-                  {
-                    conditions: '',
-                    name: '',
-                    fields: ['p99(measurements.custom.measurement)'],
-                    columns: [],
-                    aggregates: ['p99(measurements.custom.measurement)'],
-                    orderby: '-p99(measurements.custom.measurement)',
-                  },
-                ],
-              },
-            ],
-          },
-          params: {
-            widgetIndex: '0',
-          },
-          orgFeatures: [...defaultOrgFeatures, 'discover-frontend-use-events-endpoint'],
-        });
-
-        await screen.findByText('12.00hr');
-      });
-
-      it('renders custom performance metric using size units from events meta', async function () {
-        eventsMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/events/',
-          method: 'GET',
-          statusCode: 200,
-          body: {
-            meta: {
-              fields: {'p99(measurements.custom.measurement)': 'size'},
-              isMetricsData: true,
-              units: {'p99(measurements.custom.measurement)': 'kibibyte'},
-            },
-            data: [{'p99(measurements.custom.measurement)': 12}],
-          },
-        });
-
-        renderTestComponent({
-          query: {source: DashboardWidgetSource.DISCOVERV2},
-          dashboard: {
-            ...testDashboard,
-            widgets: [
-              {
-                title: 'Custom Measurement Widget',
-                interval: '1d',
-                id: '1',
-                widgetType: WidgetType.DISCOVER,
-                displayType: DisplayType.TABLE,
-                queries: [
-                  {
-                    conditions: '',
-                    name: '',
-                    fields: ['p99(measurements.custom.measurement)'],
-                    columns: [],
-                    aggregates: ['p99(measurements.custom.measurement)'],
-                    orderby: '-p99(measurements.custom.measurement)',
-                  },
-                ],
-              },
-            ],
-          },
-          params: {
-            widgetIndex: '0',
-          },
-          orgFeatures: [...defaultOrgFeatures, 'discover-frontend-use-events-endpoint'],
-        });
-
-        await screen.findByText('12.0 KiB');
-      });
-
-      it('renders custom performance metric using abyte format size units from events meta', async function () {
-        eventsMock = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/events/',
-          method: 'GET',
-          statusCode: 200,
-          body: {
-            meta: {
-              fields: {'p99(measurements.custom.measurement)': 'size'},
-              isMetricsData: true,
-              units: {'p99(measurements.custom.measurement)': 'kilobyte'},
-            },
-            data: [{'p99(measurements.custom.measurement)': 12000}],
-          },
-        });
-
-        renderTestComponent({
-          query: {source: DashboardWidgetSource.DISCOVERV2},
-          dashboard: {
-            ...testDashboard,
-            widgets: [
-              {
-                title: 'Custom Measurement Widget',
-                interval: '1d',
-                id: '1',
-                widgetType: WidgetType.DISCOVER,
-                displayType: DisplayType.TABLE,
-                queries: [
-                  {
-                    conditions: '',
-                    name: '',
-                    fields: ['p99(measurements.custom.measurement)'],
-                    columns: [],
-                    aggregates: ['p99(measurements.custom.measurement)'],
-                    orderby: '-p99(measurements.custom.measurement)',
-                  },
-                ],
-              },
-            ],
-          },
-          params: {
-            widgetIndex: '0',
-          },
-          orgFeatures: [...defaultOrgFeatures, 'discover-frontend-use-events-endpoint'],
-        });
-
-        await screen.findByText('12 MB');
-      });
-
-      it('displays custom performance metric in column select', async function () {
-        renderTestComponent({
-          query: {source: DashboardWidgetSource.DISCOVERV2},
-          dashboard: {
-            ...testDashboard,
-            widgets: [
-              {
-                title: 'Custom Measurement Widget',
-                interval: '1d',
-                id: '1',
-                widgetType: WidgetType.DISCOVER,
-                displayType: DisplayType.TABLE,
-                queries: [
-                  {
-                    conditions: '',
-                    name: '',
-                    fields: ['p99(measurements.custom.measurement)'],
-                    columns: [],
-                    aggregates: ['p99(measurements.custom.measurement)'],
-                    orderby: '-p99(measurements.custom.measurement)',
-                  },
-                ],
-              },
-            ],
-          },
-          params: {
-            widgetIndex: '0',
-          },
-          orgFeatures: [...defaultOrgFeatures, 'discover-frontend-use-events-endpoint'],
-        });
-        await screen.findByText('measurements.custom.measurement');
-      });
-
-      it('does not default to sorting by transaction when columns change', async function () {
-        renderTestComponent({
-          query: {source: DashboardWidgetSource.DISCOVERV2},
-          dashboard: {
-            ...testDashboard,
-            widgets: [
-              {
-                title: 'Custom Measurement Widget',
-                interval: '1d',
-                id: '1',
-                widgetType: WidgetType.DISCOVER,
-                displayType: DisplayType.TABLE,
-                queries: [
-                  {
-                    conditions: '',
-                    name: '',
-                    fields: [
-                      'p99(measurements.custom.measurement)',
-                      'transaction',
-                      'count()',
-                    ],
-                    columns: ['transaction'],
-                    aggregates: ['p99(measurements.custom.measurement)', 'count()'],
-                    orderby: '-p99(measurements.custom.measurement)',
-                  },
-                ],
-              },
-            ],
-          },
-          params: {
-            widgetIndex: '0',
-          },
-          orgFeatures: [...defaultOrgFeatures, 'discover-frontend-use-events-endpoint'],
-        });
-        expect(
-          await screen.findByText('p99(measurements.custom.measurement)')
-        ).toBeInTheDocument();
-        // Delete p99(measurements.custom.measurement) column
-        userEvent.click(screen.getAllByLabelText('Remove column')[0]);
-        expect(
-          screen.queryByText('p99(measurements.custom.measurement)')
-        ).not.toBeInTheDocument();
-        expect(screen.getAllByText('transaction').length).toEqual(1);
-        expect(screen.getAllByText('count()').length).toEqual(2);
       });
     });
   });
